MSMICA algorithm

This function is used to perform the MSMICA algorithm for metabolite identification using the metabolomics feature table and the KEGG database. The function takes the following inputs:

Usage

MSMICA_algorithm(
  met_raw_wide,
  metabolite_identified = NULL,
  identified_target = NULL,
  Adduct = c("M+H", "M+2Na-H", "M+Na", "M-H2O+H", "M+K", "M+2H"),
  prefix = "",
  ion_mode = "positive",
  detail = FALSE,
  adduct_correlation_r_threshold = 0.4,
  adduct_correlation_time_threshold = 5,
  isotopic_correlation_r_threshold = 0.5,
  isotopic_correlation_time_threshold = 5,
  mean_absolute_isotope_ratio_deviation_threshold = 5
)

Arguments

met_raw_wide

a metabolomics feature table in wide format with mz as the first column, time as the second column, and intensity values as the remaining columns.

metabolite_identified

a data frame containing the metabolites that have been identified in the metabolomics feature table. The first column is ion_mode (positive or negative). The second column is metabolite (the name of metabolite). The third column is KEGGID (if no KEGGID, use NA). The fourth column is adduct (the adduct form of a metabolite). The fifth column is mz and the sixth column is time. The rest of the columns are the intensity values. Default is NULL.

Adduct

a vector containing the adduct forms of the metabolites. Default is c("M+H","M+2Na-H","M+Na","M-H2O+H","M+K","M+2H") for the positive mode.

prefix

a prefix to be added to the output files. Default is "".

ion_mode

the ionization mode of the metabolomics data. Default is "positive". Other options are "negative".

detail

a logical value indicating whether to save the intermediate results as csv files. Default is FALSE. WARNING, this can create thousands of files with a lot of space. Use with caution.

adduct_correlation_r_threshold

the correlation threshold for adduct correlation analysis. Default is 0.4 (spearman correlation).

adduct_correlation_time_threshold

the retention time threshold for adduct correlation analysis. Default is 5 (seconds).

isotopic_correlation_r_threshold

the correlation threshold for isotopic correlation analysis. Default is 0.5 (spearman correlation).

isotopic_correlation_time_threshold

the retention time threshold for isotopic correlation analysis. Default is 5 (seconds).

mean_absolute_isotope_ratio_deviation_threshold

the threshold for the mean absolute isotope ratio deviation between predicted isotopic relative abundance and actual isotopic relative abundance compared to the primary adduct without heavy isotope. Default is 5 (percent).

identified_target:

a vector containing the KEGGID that are used as additional starting points for the MSMICA algorithm. This can be used when users just want to use specific identified metabolites as the starting point for the MSMICA algorithm. Default is NULL.