target_metabolite_search

This function searches for target metabolites in a reference library and a feature table.

Usage

target_metabolite_search(
  reference_library,
  feature_table,
  use_retention_time = TRUE,
  mz_threshold = 10,
  time_threshold = 30
)

Arguments

reference_library

A reference library table. The reference mz should be called "reference_mz" and the reference retention time should be called "reference_time".

feature_table

A feature table. The mz should be called "mz" and the time should be called "time".

use_retention_time

A logical value indicating whether to use retention time for matching. Default is TRUE.

mz_threshold

The m/z threshold for matching features between two data sets. Default is 10 ppm.

time_threshold

The retention time threshold for matching features between two data sets. Default is 30 seconds. If you use minutes, please convert the unit appropriately.

Value

A data frame with the overlapping features, including `mz_ppm_difference` and, when retention-time matching is used, `time_difference`.